Refactored code, added add task button to page
