+++ front-end/src/App.css
@@ -7,13 +7,7 @@
   position: relative;
   margin: 0 auto;
   height: 100vh;
-  width: 100%;
-}
-
-@media (min-width: 768px) {
-  .App {
-    max-width: 428px;
-  }
+  width: var(--max-width);
 }
