+++ front-end/src/components/Task_Component.jsx
@@ -18,12 +18,12 @@ import Checkbox from "@mui/material/Checkbox";
  * @param {string} assigned person task is assigned to, if any
  * @returns The task component as highlighted below
  */
-const TaskComponent = ({id, title, room, assigned, completed, SelectHandler}) => {
+const TaskComponent = ({id, title, room, assigned, completed, SelectHandler, enableCheckbox}) => {
   const [isChecked, setIsChecked] = React.useState(false);
 
   return (
     <div className={"task_box_comp "+(completed?"complete":"incomplete")}>
-      <Checkbox className="checkbox" checked={isChecked && !completed} onChange={e => {setIsChecked(e.target.checked); SelectHandler(id, e.target.checked && !completed)}}/>
+      {enableCheckbox && (<Checkbox className="checkbox" checked={isChecked && !completed} onChange={e => {setIsChecked(e.target.checked); SelectHandler(id, e.target.checked && !completed)}}/>)}
       <div className={"task_component "+(isChecked && !completed ?"selected":"deselected")}>
         <div className="task_information">
           <div className="task_name">
