+++ front-end/src/css/Footer.css
@@ -7,12 +7,5 @@
   justify-content: space-between;
   height: var(--footer-height);
   background-color: white;
-  width: 100%;
-}
-
-
-@media (min-width: 768px) {
-  .footer {
-    max-width: 428px;
-  }
+  width: var(--max-width);
 }
