+++ front-end/src/css/Header.css
@@ -7,19 +7,13 @@
   justify-content: space-between;
   height: var(--header-height);
   background-color: #f2cc8f;
-  width: 100%;
+  width: var(--max-width);
 }
 
 .header + * {
   padding-top: var(--header-height);
   padding-bottom: var(--footer-height);
   height: calc(100vh - var(--header-height) - var(--footer-height));
-}
-
-@media (min-width: 768px) {
-  .header {
-    max-width: 428px;
-  }
 }
 
 .headerLogo {
