+++ front-end/src/css/IndividualTask.css
@@ -1,4 +1,4 @@
 .taskFormContainer {
-    height: 100vh;
+    width: 100%;
     background-color: #f4f1de;
 }
