+++ front-end/src/css/task_list_component.css
@@ -1,0 +1,64 @@
+
+.task_box {
+  position: relative;
+  text-align: left;
+  padding-top: 3rem;
+  padding-bottom: 4rem;
+  width: 100%;
+}
+
+.loading_icon {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
+
+.task_box > .transparent_box {
+  position: fixed;
+  z-index: 100;
+  bottom: var(--footer-height);
+  left: 50%;
+  width: var(--max-width);
+  height: 10vh;
+  transform: translate(-50%, 0);
+  background-image: linear-gradient(
+    to bottom,
+    rgba(255, 255, 255, 0),
+    rgba(255, 255, 255, 1) 90%
+  );
+}
+
+.task_box > .left_button {
+  position: fixed;
+  z-index: 9999;
+  bottom: var(--footer-height);
+  left: calc(50% - var(--max-width) * 0.2);
+  width: 12rem;
+  transform: translate(-50%, -50%);
+  border-radius: 15em;
+}
+
+.task_box > .center_button {
+  position: fixed;
+  z-index: 9999;
+  bottom: var(--footer-height);
+  left: 50%;
+  width: 12rem;
+  transform: translate(-50%, -50%);
+  border-radius: 15em;
+}
+
+.task_box .line {
+  height: 0.2rem;
+  background: repeating-linear-gradient(90deg, lightgray 0 2rem, #0000 0 4rem);
+}
+
+.task_box .day_title {
+  display: block;
+  background-size: 100px;
+  padding-bottom: 2rem;
+  padding-left: 1rem;
+  text-align: left;
+  margin: 0.8rem 0rem;
+}
