+++ front-end/src/css/tasks.css
@@ -1,61 +1,9 @@
-/* .tasks-page-content {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  padding-top: 2rem;
-  padding-bottom: 5rem;
-} */
-
-.task_box {
-  position: relative;
-  text-align: left;
-  padding-top: 3rem;
-  padding-bottom: 4rem;
-  width: 100%;
-}
-
-.loading_icon {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-.task_box > .transparent_box {
-  position: fixed;
-  z-index: 100;
-  bottom: var(--footer-height);
-  left: 50%;
-  width: 100vw;
-  height: 10vh;
-  transform: translate(-50%, 0);
-  background-image: linear-gradient(
-    to bottom,
-    rgba(255, 255, 255, 0),
-    rgba(255, 255, 255, 1) 90%
-  );
-}
-
-.task_box > Button {
+.tasks_page > .add_tasks_button {
   position: fixed;
   z-index: 9999;
   bottom: var(--footer-height);
-  left: 50%;
-  width: 12rem;
+  left: calc(50% + var(--max-width) * 0.28);
+  border-radius: 15em;
   transform: translate(-50%, -50%);
 }
 
-.task_box .line {
-  height: 0.2rem;
-  background: repeating-linear-gradient(90deg, lightgray 0 2rem, #0000 0 4rem);
-}
-
-.task_box .day_title {
-  display: block;
-  background-size: 100px;
-  padding-bottom: 2rem;
-  padding-left: 1rem;
-  text-align: left;
-  margin: 0.8rem 0rem;
-}
