+++ front-end/src/pages/IndividualTask.jsx
@@ -108,7 +108,6 @@ function IndividualTask(props) {
                     </Box>
                 </ThemeProvider>
             </form>
-
             <Footer />
         </>
     )
